Refactor test class to extend test base class
